+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The last regular expression in the tokenizer has been replaced with
   a manually written matcher.  This further improves performance of a
   portscan and brings it down to ~35 s.
+- portscan: The progress report has been enabled by default with the
+  interval shortened to 1 s when `stderr` is a TTY.  The output
+  is also kept on only one line in that case.  `--progress=0`
+  disables this.
 
 ## [1.0.0] - 2021-09-01